+++ lib/bugno.rb
@@ -1,13 +1,16 @@
 # frozen_string_literal: true
 
 require 'bugno/logger'
-require 'bugno/generator/bugno_generator'
 require 'bugno/configuration'
-require 'bugno/railtie'
+require 'bugno/event'
+
+if defined?(Rails)
+  require 'bugno/railtie'
+  require 'bugno/generator/bugno_generator'
+end
 
 module Bugno
-  class Error < StandardError
-  end
+  class Error < StandardError; end
 
   class << self
     attr_accessor :configuration
@@ -16,12 +19,8 @@ module Bugno
       @configuration ||= Configuration.new
     end
 
-    def reset
-      @configuration = Configuration.new
-    end
-
     def configured?
-      configuration.api_key
+      !!configuration.api_key
     end
 
     def configure